Move internal packages to internal/ folder
